Move store logic to its own separate file
